+++ JsonFlatFileDataStore/IDocumentCollection.cs
@@ -75,6 +75,15 @@ namespace JsonFlatFileDataStore
         bool ReplaceOne(Predicate<T> filter, T item, bool upsert = false);
 
         /// <summary>
+        /// Replace the item matching the id
+        /// </summary>
+        /// <param name="id">The item mathcing the id-value will be replaced</param>
+        /// <param name="item">New content</param>
+        /// <param name="upsert">Will item be inserted if not found</param>
+        /// <returns>true if item found for replacement</returns>
+        bool ReplaceOne(dynamic id, T item, bool upsert = false);
+
+        /// <summary>
         /// Replace the first item matching the filter
         /// </summary>
         /// <param name="filter">First item matching the predicate will be replaced</param>
@@ -82,6 +91,15 @@ namespace JsonFlatFileDataStore
         /// <param name="upsert">Will item be inserted if not found</param>
         /// <returns>true if item found for replacement</returns>
         Task<bool> ReplaceOneAsync(Predicate<T> filter, T item, bool upsert = false);
+
+        /// <summary>
+        /// Replace the item matching the id
+        /// </summary>
+        /// <param name="id">The item matching the id-value will be replaced</param>
+        /// <param name="item">New content</param>
+        /// <param name="upsert">Will item be inserted if not found</param>
+        /// <returns>true if item found for replacement</returns>
+        Task<bool> ReplaceOneAsync(dynamic id, T item, bool upsert = false);
 
         /// <summary>
         /// Replace all items matching the filter
@@ -108,12 +126,28 @@ namespace JsonFlatFileDataStore
         bool UpdateOne(Predicate<T> filter, dynamic item);
 
         /// <summary>
+        /// Update the item matching the id
+        /// </summary>
+        /// <param name="id">The item matching the id-value will be replaced</param>
+        /// <param name="item">New content</param>
+        /// <returns>true if item found for update</returns>
+        bool UpdateOne(dynamic id, dynamic item);
+
+        /// <summary>
         /// Update the first item matching the filter
         /// </summary>
         /// <param name="filter">First item matching the predicate will be replaced</param>
         /// <param name="item">New content</param>
         /// <returns>true if item found for update</returns>
         Task<bool> UpdateOneAsync(Predicate<T> filter, dynamic item);
+
+        /// <summary>
+        /// Update the item matching the id
+        /// </summary>
+        /// <param name="id">The item matching the id-value will be replaced</param>
+        /// <param name="item">New content</param>
+        /// <returns>true if item found for update</returns>
+        Task<bool> UpdateOneAsync(dynamic id, dynamic item);
 
         /// <summary>
         /// Update all items matching the filter
@@ -139,11 +173,25 @@ namespace JsonFlatFileDataStore
         bool DeleteOne(Predicate<T> filter);
 
         /// <summary>
+        /// Delete the item matching the id
+        /// </summary>
+        /// <param name="id">The item matching the id-value will be deleted</param>
+        /// <returns>true if item found for deletion</returns>
+        bool DeleteOne(dynamic id);
+
+        /// <summary>
         /// Delete first item matching the filter
         /// </summary>
         /// <param name="filter">First item matching the predicate will be deleted</param>
         /// <returns>true if item found for deletion</returns>
         Task<bool> DeleteOneAsync(Predicate<T> filter);
+
+        /// <summary>
+        /// Delete the item matching the id
+        /// </summary>
+        /// <param name="id">The item matching the id-vvalue will be deleted</param>
+        /// <returns>true if item found for deletion</returns>
+        Task<bool> DeleteOneAsync(dynamic id);
 
         /// <summary>
         /// Delete all items matching the filter